+++ src/main/java/com/github/euonmyoji/epicbanitem/util/file/ObservableDirectory.java
@@ -1,4 +1,4 @@
-package com.github.euonmyoji.epicbanitem;
+package com.github.euonmyoji.epicbanitem.util.file;
 
 import java.io.Closeable;
 import java.io.IOException;
