+++ src/main/java/com/github/euonmyoji/epicbanitem/util/file/ObservableFileService.java
@@ -1,4 +1,4 @@
-package com.github.euonmyoji.epicbanitem;
+package com.github.euonmyoji.epicbanitem.util.file;
 
 import com.google.inject.ImplementedBy;
