+++ src/main/java/com/github/euonmyoji/epicbanitem/util/file/Savable.java
@@ -1,8 +1,7 @@
-package com.github.euonmyoji.epicbanitem;
+package com.github.euonmyoji.epicbanitem.util.file;
 
 import java.io.IOException;
 
 public interface Savable {
-
     void save() throws IOException;
 }
